lexopt

Enum Error

pub enum Error {
    MissingValue {
        option: Option<String>,
    },
    UnexpectedOption(String),
    UnexpectedArgument(OsString),
    UnexpectedValue {
        option: String,
        value: OsString,
    },
    ParsingFailed {
        value: String,
        error: Box<dyn Error + Send + Sync + 'static>,
    },
    NonUnicodeValue(OsString),
    Custom(Box<dyn Error + Send + Sync + 'static>),
}

An error during argument parsing.

This implements From<String> and From<&str>, for easy ad-hoc error messages.

Variants

MissingValue

An option argument was expected but was not found.

Fields

option: Option<String>

The most recently emitted option.

UnexpectedOption(String)

An unexpected option was found.

UnexpectedArgument(OsString)

A positional argument was found when none was expected.

UnexpectedValue

An option had a value when none was expected.

Fields

option: String

The option.

value: OsString

The value.

ParsingFailed

Parsing a value failed. Returned by methods on ValueExt.

Fields

value: String

The string that failed to parse.

error: Box<dyn Error + Send + Sync + 'static>

The error returned while parsing.

NonUnicodeValue(OsString)

A value was found that was not valid unicode.

This can be returned by the methods on ValueExt.

Custom(Box<dyn Error + Send + Sync + 'static>)

For custom error messages in application code.